link_available_for_payment is no longer Terra specific, it is moved to relay and is updated here accordingly
