+++ pkg/monitoring/metrics.go
@@ -10,7 +10,6 @@ import (
 type Metrics interface {
 	SetProxyAnswersRaw(answer float64, proxyContractAddress, feedID, chainID, contractStatus, contractType, feedName, feedPath, networkID, networkName string)
 	SetProxyAnswers(answer float64, proxyContractAddress, feedID, chainID, contractStatus, contractType, feedName, feedPath, networkID, networkName string)
-	SetLinkAvailableForPayment(amount float64, feedID, chainID, contractStatus, contractType, feedName, feedPath, networkID, networkName string)
 	Cleanup(proxyContractAddress, feedID, chainID, contractStatus, contractType, feedName, feedPath, networkID, networkName string)
 }
 
@@ -28,13 +27,6 @@ var (
 			Help: "Reports the latest answer from the proxy contract divided by the feed's multiplier parameter.",
 		},
 		[]string{"proxy_contract_address", "feed_id", "chain_id", "contract_status", "contract_type", "feed_name", "feed_path", "network_id", "network_name"},
-	)
-	linkAvailableForPayment = promauto.NewGaugeVec(
-		prometheus.GaugeOpts{
-			Name: "link_available_for_payments",
-			Help: "Reports the amount of link the contract can use to make payments to node operators. This may be different from the LINK balance of the contract since that can contain debt",
-		},
-		[]string{"feed_id", "chain_id", "contract_status", "contract_type", "feed_name", "feed_path", "network_id", "network_name"},
 	)
 )
 
@@ -75,19 +67,6 @@ func (d *defaultMetrics) SetProxyAnswers(answer float64, proxyContractAddress, f
 	}).Set(answer)
 }
 
-func (d *defaultMetrics) SetLinkAvailableForPayment(amount float64, feedID, chainID, contractStatus, contractType, feedName, feedPath, networkID, networkName string) {
-	linkAvailableForPayment.With(prometheus.Labels{
-		"feed_id":         feedID,
-		"chain_id":        chainID,
-		"contract_status": contractStatus,
-		"contract_type":   contractType,
-		"feed_name":       feedName,
-		"feed_path":       feedPath,
-		"network_id":      networkID,
-		"network_name":    networkName,
-	}).Set(amount)
-}
-
 func (d *defaultMetrics) Cleanup(
 	proxyContractAddress, feedID, chainID, contractStatus, contractType string,
 	feedName, feedPath, networkID, networkName string,
@@ -108,18 +87,5 @@ func (d *defaultMetrics) Cleanup(
 	}
 	if !proxyAnswers.Delete(labels) {
 		d.log.Errorw("failed to delete metric", "name", "proxy_answers", "labels", labels)
-	}
-	linkLeftLabels := prometheus.Labels{
-		"feed_id":         feedID,
-		"chain_id":        chainID,
-		"contract_status": contractStatus,
-		"contract_type":   contractType,
-		"feed_name":       feedName,
-		"feed_path":       feedPath,
-		"network_id":      networkID,
-		"network_name":    networkName,
-	}
-	if !linkAvailableForPayment.Delete(linkLeftLabels) {
-		d.log.Errorw("failed to delete metric", "name", "link_available_for_payment", "labels", linkLeftLabels)
 	}
 }
