+++ pkg/monitoring/source_proxy.go
@@ -5,18 +5,13 @@ import (
 	"encoding/json"
 	"fmt"
 	"math/big"
-	"sync"
 
 	relayMonitoring "github.com/smartcontractkit/chainlink-relay/pkg/monitoring"
-	"go.uber.org/multierr"
 )
 
 // ProxyData is a subset of the data returned by the Terra feed proxy contract's "latest_round_data" method.
 type ProxyData struct {
 	Answer *big.Int
-	// This value is not taken from the proxy contract but from the backend contract.
-	// It's added here for convenience.
-	LinkAvailableForPayment *big.Int
 }
 
 // NewProxySourceFactory does wisott.
@@ -65,35 +60,11 @@ func (p *proxySource) Fetch(ctx context.Context) (interface{}, error) {
 		p.log.Debugw("skipping fetch because no proxy contract is configured", "feed", p.terraFeedConfig.ContractAddressBech32)
 		return nil, relayMonitoring.ErrNoUpdate
 	}
-	proxyData := ProxyData{}
-	var proxyErr error
-	proxyDataMu := &sync.Mutex{}
-	wg := &sync.WaitGroup{}
-	wg.Add(2)
-	go func() {
-		defer wg.Done()
-		answer, err := p.fetchLatestRoundFromProxy(ctx)
-		proxyDataMu.Lock()
-		defer proxyDataMu.Unlock()
-		if err != nil {
-			proxyErr = multierr.Combine(proxyErr, err)
-		} else {
-			proxyData.Answer = answer
-		}
-	}()
-	go func() {
-		defer wg.Done()
-		amount, err := p.fetchLinkAvailableForPayment(ctx)
-		proxyDataMu.Lock()
-		defer proxyDataMu.Unlock()
-		if err != nil {
-			proxyErr = multierr.Combine(proxyErr, err)
-		} else {
-			proxyData.LinkAvailableForPayment = amount
-		}
-	}()
-	wg.Wait()
-	return proxyData, proxyErr
+	answer, err := p.fetchLatestRoundFromProxy(ctx)
+	if err != nil {
+		return nil, err
+	}
+	return ProxyData{answer}, nil
 }
 
 // latestRoundDataRes corresponds to a subset of the Round type in the proxy contract.
@@ -119,28 +90,4 @@ func (p *proxySource) fetchLatestRoundFromProxy(ctx context.Context) (*big.Int, 
 		return nil, fmt.Errorf("failed to parse proxy answer '%s' into a big.Int", latestRoundData.Answer)
 	}
 	return answer, nil
-}
-
-type linkAvailableForPaymentRes struct {
-	Amount string `json:"amount,omitempty"`
-}
-
-func (p *proxySource) fetchLinkAvailableForPayment(ctx context.Context) (*big.Int, error) {
-	res, err := p.client.ContractStore(
-		ctx,
-		p.terraFeedConfig.ContractAddress,
-		[]byte(`"link_available_for_payment"`),
-	)
-	if err != nil {
-		return nil, fmt.Errorf("failed to read link_available_for_payment from the proxy contract: %w", err)
-	}
-	linkAvailableForPayment := linkAvailableForPaymentRes{}
-	if err := json.Unmarshal(res, &linkAvailableForPayment); err != nil {
-		return nil, fmt.Errorf("failed to unmarshal link available data from the response '%s': %w", string(res), err)
-	}
-	amount, success := new(big.Int).SetString(linkAvailableForPayment.Amount, 10)
-	if !success {
-		return nil, fmt.Errorf("failed to parse amount of link available for payment from string '%s' into a big.Int", linkAvailableForPayment.Amount)
-	}
-	return amount, nil
 }
